Add DimensionSelector id -> X caches.
Adds the method DimensionSelectorUtils.cacheIfPossible to aid with
creation of caches for evaluating functions of dimension dictionary IDs.
Uses an array cache when possible, an LRU cache when it seems like it
might be a good idea, and skips caching otherwise.

The caches are used in three places in this patch:

- Cardinality aggregator (caches hashes)
- Single-dimension-input expression selector (caches expression results)
- Generic expression selector (caches name for any dimension inputs)

Benchmarking showed that the LRU cache could introduce unwanted overhead
for dimensions with few repeating values, so this patch implements two
mitigations:

- Skip caching completely if cardinality is >90% of the row count.
- After iterating through a few multiples of the cache size, examine
  the cache hit rate and decide to either freeze it (the cache becomes
  read-only) or ignore it (the cache is cleared and future operations
  are uncached).

The caches aim to be <1MB per column.